chore: Update libc and criterion crates to latest

+++ bench/src/bench.rs
@@ -749,7 +749,7 @@ fn define(
     corpus: &[u8],
     bench: impl FnMut(&mut Bencher) + 'static,
 ) {
-    let tput = Throughput::Bytes(corpus.len() as u32);
+    let tput = Throughput::Bytes(corpus.len() as u64);
     let benchmark = Benchmark::new(bench_name, bench).throughput(tput);
     c.bench(group_name, benchmark);
 }
